Add food strength customization, refactor config reading

+++ src/size.rs
@@ -1,5 +1,5 @@
 #[derive(Copy, Clone, Debug, PartialEq)]
 pub struct Size {
-    pub width: usize,
-    pub height: usize,
+    pub width: u16,
+    pub height: u16,
 }
